rg-6: * Changed functionality of component to navbar.
