--- frontend/src/bundles/common/components/tabs/tab-item.tsx
@@ -1,12 +1,0 @@
-import React, { type ReactNode } from 'react';
-
-type TabItemProperties = {
-    children: ReactNode;
-    title: string;
-};
-
-const TabItem: React.FC<TabItemProperties> = ({ children }) => {
-    return <div>{children}</div>;
-};
-
-export { TabItem };
